all done

+++ 0x00-hello_world/6-size.c
@@ -12,14 +12,14 @@ int main(void)
 char letters;
 int myNum;
 long longNum;
-double doubleNum;
+long long doubleNum;
 float money;
 
-printf("Size of the char is %d\n", (unsigned)sizeof(letters));
-printf("Size of the int is %d\n", (unsigned)sizeof(myNum));
-printf("Size of the long is %d\n", (unsigned)sizeof(longNum));
-printf("Size of the long long is %d\n", (unsigned)sizeof(doubleNum));
-printf("Size of the float is %d\n", (unsigned)sizeof(money));
+printf("Size of the char is %lu\n", (unsigned)sizeof(letters));
+printf("Size of the int is %lu\n", (unsigned)sizeof(myNum));
+printf("Size of the long is %lu\n", (unsigned)sizeof(longNum));
+printf("Size of the long long is %lu\n", (unsigned)sizeof(doubleNum));
+printf("Size of the float is %lu\n", (unsigned)sizeof(money));
 
 return (0);
 }
